maint: broke out linked account connections tests
